started splitting up the notebooks and cleaning code; need to work on adding Aaron's terms and compare to bSDD
