+++ .gitignore
@@ -1,5 +1,5 @@
 # SPaR.txt to avoid uploading a large model file
-SPaRtxt/
+SPaR.txt/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
